+++ src/plugins/core/template/core_template_api.c
@@ -56,7 +56,7 @@ static int template_send(cci_connection_t *connection,
                          void *msg_ptr, uint32_t msg_len,
                          void *context, int flags);
 static int template_sendv(cci_connection_t *connection,
-                          struct iovec *data, uint8_t iovcnt,
+                          struct iovec *data, uint32_t iovcnt,
                           void *context, int flags);
 static int template_rma_register(cci_endpoint_t *endpoint,
                                  cci_connection_t *connection,
@@ -284,7 +284,7 @@ static int template_send(cci_connection_t *connection,
 
 
 static int template_sendv(cci_connection_t *connection,
-                          struct iovec *data, uint8_t iovcnt,
+                          struct iovec *data, uint32_t iovcnt,
                           void *context, int flags)
 {
     printf("In template_sendv\n");
